voice-service/requirements.txt
# Core API and web server
fastapi>=0.95.0
uvicorn>=0.21.1
pydantic>=2.0.0
python-multipart>=0.0.6
starlette>=0.27.0

# Speech-to-Text and Text-to-Speech
torch==2.5.1
torchaudio==2.5.1
numpy==1.22.0
soundfile>=0.12.1
openai-whisper>=20230314
TTS>=0.17.0
phonemizer>=3.2.1
espeakng>=1.0.0
scipy~=1.11.2
matplotlib~=3.7.0
librosa~=0.10.0
numba~=0.57.0

# Audio and file utilities
ffmpeg-python>=0.2.0

# Caching and rate limiting
redis>=4.5.4
fastapi-limiter>=0.1.5

# Logging and environment
python-json-logger>=2.0.7
python-dotenv

# HTTP and API
requests>=2.28.2

# ElevenLabs
elevenlabs>=1.0.0

# Google Cloud Speech
google-cloud-speech>=2.21.0

# Azure Speech
azure-cognitiveservices-speech>=1.34.1

# AWS SDK
boto3>=1.34.0
s3transfer>=0.7.0

# Language detection
langdetect

# Parler TTS
git+https://github.com/huggingface/parler-tts.git

# LLM and AI Services
openai>=1.0.0
anthropic>=0.7.0
tiktoken>=0.5.0

# WebSocket support
websockets>=11.0.0

# Additional utilities for commoncode
aiohttp>=3.8.0
asyncio-mqtt>=0.13.0

